refactoring versions
